4/5/21 add SQLite database functionality, though i have yet to see it work due to issues with close() somewhere... there's an entirely new document that keeps track of database functions and added a delete button that toggles visibility with whether an edit flag is sent to addClient. Tried my hand at Firebase, though Gradle quickly shot that effort down with request to accept terms of service that they didn't want to accept...

+++ app/src/main/java/org/wit/example/beautycentral/models/ClientMemStore.kt
@@ -3,9 +3,9 @@ package org.wit.example.beautycentral.models
 import org.jetbrains.anko.AnkoLogger
 import org.jetbrains.anko.info
 
-var lastId=0L
+var lastId=0
 
-internal fun getId():Long{
+internal fun getId(): Int {
     return lastId++
 }
